Remove unnecessary `mitmTools` namespace
